requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.1.0
gunicorn==23.0.0
num2words==0.5.14
numpy==2.0.2
pyaudio==0.2.14
pvporcupine==3.0.5
vosk==0.3.45
word2number==1.1
